+++ src/models/Settings.php
@@ -1,31 +1,11 @@
 <?php
-/**
- * Shortcut plugin for Craft CMS 3.x
- *
- * Simple URL shortening
- *
- * @link      https://superbig.co
- * @copyright Copyright (c) 2017 Superbig
- */
+namespace verbb\shortcut\models;
 
-namespace superbig\shortcut\models;
-
-use craft\helpers\UrlHelper;
-use craft\services\Elements;
-use superbig\shortcut\Shortcut;
-
-use Craft;
 use craft\base\Model;
-use yii\base\Exception;
 
-/**
- * @author    Superbig
- * @package   Shortcut
- * @since     1.0.0
- */
 class Settings extends Model
 {
-    // Public Properties
+    // Properties
     // =========================================================================
 
     /**
@@ -43,18 +23,17 @@ class Settings extends Model
      */
     public $customDomain = '';
 
+
     // Public Methods
     // =========================================================================
 
-    /**
-     * @inheritdoc
-     */
-    public function rules ()
+    public function defineRules(): array
     {
-        return [
-            [ 'urlSegment', 'string' ],
-            [ 'hideUrlSegment', 'bool' ],
-            [ 'customDomain', 'string' ],
-        ];
+        $rules = parent::defineRules();
+
+        $rules[] = [['urlSegment', 'customDomain'], 'string'];
+        $rules[] = [['hideUrlSegment'], 'bool'];
+
+        return $rules;
     }
 }