+++ src/records/Shortcut.php
@@ -1,0 +1,15 @@
+<?php
+namespace verbb\shortcut\records;
+
+use craft\db\ActiveRecord;
+
+class Shortcut extends ActiveRecord
+{
+    // Static Methods
+    // =========================================================================
+
+    public static function tableName(): string
+    {
+        return '{{%shortcut_shortcuts}}';
+    }
+}