+++ src/translations/en/shortcut.php
@@ -1,18 +1,5 @@
 <?php
-/**
- * Shortcut plugin for Craft CMS 3.x
- *
- * Simple URL shortening
- *
- * @link      https://superbig.co
- * @copyright Copyright (c) 2017 Superbig
- */
 
-/**
- * @author    Superbig
- * @package   Shortcut
- * @since     1.0.0
- */
 return [
-    'Shortcut plugin loaded' => 'Shortcut plugin loaded',
+    'Could not find the url for element {id}' => 'Could not find the url for element {id}',
 ];