+++ CHANGELOG.md
@@ -1,22 +1,30 @@
-# Shortcut Changelog
+# Changelog
 
-All notable changes to this project will be documented in this file.
+## 3.0.0 - 2022-06
 
-The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
+> {note} The plugin’s package name has changed to `verbb/shortcut`. Shortcut will need be updated to 3.0 from a terminal, by running `composer require verbb/shortcut && composer remove superbig/craft3-shortcut`.
+
+### Changed
+- Migration to `verbb/shortcut`.
+- Now requires Craft 3.7+.
 
 ## 2.0.3 - 2019-09-03
+
 ### Fixed
 - Removed stray `var_dump`
 
 ## 2.0.2 - 2018-10-18
+
 ### Changed
 - Added proper migration for upgrading from Craft 2 to Craft 3
 - Removed Hashids dependency
 
 ## 2.0.1 - 2018-06-01
+
 ### Fixed
 - Fixed create method for url's
 
 ## 2.0.0 - 2017-11-14
+
 ### Added
 - Initial release